make validate and transform aliases to the same hook
Since validate and transform can do the same things, let's make them aliases for the same hook. I think it is still good to have the names to communicate the intent of the logic. You can now only define one or the other.

The code becomes much simpler now since we don't have to deal with both validate and transform code.

createLogic will now throw an error if you define both in the same logic.

Internally this hook will be called validtrans for which validate and transform are aliases.