createLogicMiddleware errors on non-array, add more jsdoc
throw an error when a non-array item is passed into createLogicMiddleware, this prevents users from thinking they can pass in the logic items as separate parameters when instead they should be passing them in as an array.

Catching this early will help prevent confusion.

add jsdoc for addLogic and replaceLogic methods