search-async-fetch example, also change single-file to search
Since building a live search is great example of using the power of redux-logic, I have added a search-async-fetch example which uses debouncing and the take latest features.

Also I converted the single-file example to search as well so you can see everything in a single file.

Finally I added a jsfiddle example which demonstrates search as well.