document name, whenComplete(), and monitor$
optional name for each logic

mw.whenComplete(fn) - calls fn and returns promise when all in-flight processing has been completed. Useful for async loading on the server. Dispatch actions then call whenComplete to know when they are all done and it is safe to serialize state.getState()

mw.monitor$ an observable which can be used to monitor the internals of redux-logic. Can be used for logging or to develop dev tools.