be sure to console.error errs in catch since could be render err
Update examples to all console.error(err) in any catch blocks
since it could be hiding render errors. If there is a render err
from a previous dispatch then the components might not be able
to display the error, so be sure to log to console.

When using processOptions or observables things work properly since the dispatch is outside of the try catch blocks.