make the examples use the full multiple-dispatch signature
Since it is easy for new users to become confused if they take an example and don't understand the single/multiple dispatch modes and if they try to do multiple dispatches then they are confused about why it didn't work.

It would be better to just default to the multiple-dispatch signature even though it is more verbose, it is safer for people to use in any manner they choose.