switch examples to use multi-dispatch process with done and https
Switched the examples to using the multi-dispatch mode with done so in case people cut and paste code and then try to do multiple dispatching then it will work properly. They wouldn't understand that without the done, dispatch is single dispatch. Also it is confusing with single-dispatch to need to call dispatch even if nothing to dispatch.

So it is safer to just use the multidispatch mode for examples. We may deprecate the single-dispatch mode in the future to prevent confusion.

Also switched examples to using https since reqres.in now supports it.