clean up dispatch and return handling, add more tests
Refine the dispatch and return handling especially around the areas of error handling and use of falsey values.

Clarify the details of how values are used in the API and also what affect successType and failType properties have. Specifically call out that if the action creator fn form is used, it can return falsey value like undefined if it doesn't wish to dispatch anything.

Clarify the flow for errors both returned, rejected, thrown, or from observables. If failType is not set, then check Error for a type and if found it can be dispatched as is, otherwise wrap in an UNHANDLED_LOGIC_ERROR action and provide the error value as the payload.

Add more tests to verify the operation of all these situations and when errors are thrown in the code.