better custom rxbuild checking, add 2017 to license
Since we use a custom Rx build, check for the props on load of the files so we can instantly verify that these exist. This needs to be done in isolation since other tests load the full Rx for convenience.

Also run the code through a typical run to also try to catch anything that was not specified up top.

add 2017 to license