add imports to examples, add warnTimeout, future depr
add imports to README examples.

add `warnTimeout` option to `createLogic`. Used only in development environment. It defaults to 60000 ms (one minute) and it will write to console.error if logic has not completed by the time this fires. This will help to find code that is not completing by calling `done()`. If your logic truly does not end (like in a web socket scenario) you can set the warnTimeout to `0` and it will not run.

Also note in the docs that in a future version dispatchMultiple will no longer be needed once single-dispatch is removed. Single-dispatch is deprecated in this version.