console.error reducer/dispatch errors
If an error occurred in the reducer during dispatch then RxJS was
capturing the error and there was no indication of what went wrong.

This fix will log the error to console.error and also indicates in
the monitor$ with a new op: 'dispatchError' that an error occurred.

Errors in reducer are somewhat difficult since we don't know that
we can safely dispatch anything additional if a reducer is erroring,
thus we log to console.error as the best mechanism.