prevent hiding of errors occuring during a mw next call
If an error is thrown during a middleware next call (in the downstream
middleware, reducer, or renderer) then it could possibly be hidden,
so this change logs it to console.error as well as the monitor$ to
help with debugging.

Because we are now catching errors during the next call, we also
pick up those that happen in dispatch as well, so we don't need
special handling for those, they all get picked up by nextError.

Document the additional op when an error is detected (nextError).